Store emulation is not restored correctly
If I save a product in admin area and the product is in two stores especially with different languages, I've got problems with the store emulation. There is a call to start emulation for each store, but there is only one restoring. So the following happens:

1. I'm in admin store
2. Emulating german store for loading product data
3. Emulating english store for loading product data
4. Stopping emulation and setting it to german store, because the initial environment in step 3 is the german store.
5. The admin area is now in a wrong state and I have to login again.

This problem should be fixed with line 421. But the translation is also not restored correctly. So I switched the position of the locale setting in method _emulateStore, because otherwise when starting the emulation, a wrong initial locale would be set.